+++ packages/plugins/typescript/resolvers/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @graphql-codegen/typescript-resolvers
 
+## 4.4.2
+
+### Patch Changes
+
+- [#10230](https://github.com/dotansimha/graphql-code-generator/pull/10230) [`60dd72f`](https://github.com/dotansimha/graphql-code-generator/commit/60dd72fb103fd7fd70b4e1def98da29588865517) Thanks [@eddeee888](https://github.com/eddeee888)! - Fix generateInternalResolversIfNeeded.\_\_resolveReference making the resolver required
+
+  `__resolveReference`'s default behaviour when not declared is to pass the ref to subsequent resolvers i.e. becoming the `parent`. So, it means we don't have to make this resolver required.
+
+  This patch makes `__resolveReference` optional when `generateInternalResolversIfNeeded.__resolveReference` is set to true.
+
+- Updated dependencies [[`60dd72f`](https://github.com/dotansimha/graphql-code-generator/commit/60dd72fb103fd7fd70b4e1def98da29588865517)]:
+  - @graphql-codegen/visitor-plugin-common@5.6.1
+  - @graphql-codegen/typescript@4.1.3
+
 ## 4.4.1
 
 ### Patch Changes